requirements.txt
fastapi
uvicorn
torch
torchvision
transformers
pillow
numpy
python-multipart
scipy
gdown